+++ components/integralui.menu.js
@@ -1,17 +1,3 @@
-/*
-  filename: integralui.menu.js
-  version : 0.7.524 BETA
-  Copyright © 2016-2017 Lidor Systems. All rights reserved.
-
-  This file is part of the "IntegralUI Web" Library. 
-                                                                   
-  The contents of this file are subject to the IntegralUI Web License, and may not be used except in compliance with the License.
-  A copy of the License should have been installed in the product's root installation directory or it can be found at
-  http://www.lidorsystems.com/products/web/iui-web-license-agreement.pdf.
-                                                            
-  This SOFTWARE is provided "AS IS", WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the specific language 
-  governing rights and limitations under the License. Any infringement will be prosecuted under applicable laws.                           
-*/
 var __extends=this&&this.__extends||function(c,a){function d(){this.constructor=c}for(var b in a)a.hasOwnProperty(b)&&(c[b]=a[b]);c.prototype=null===a?Object.create(a):(d.prototype=a.prototype,new d)},__decorate=this&&this.__decorate||function(c,a,d,b){var f=arguments.length,e=3>f?a:null===b?b=Object.getOwnPropertyDescriptor(a,d):b,g;if("object"===typeof Reflect&&"function"===typeof Reflect.decorate)e=Reflect.decorate(c,a,d,b);else for(var h=c.length-1;0<=h;h--)if(g=c[h])e=(3>f?g(e):3<f?g(a,d,e):
 g(a,d))||e;return 3<f&&e&&Object.defineProperty(a,d,e),e},__metadata=this&&this.__metadata||function(c,a){if("object"===typeof Reflect&&"function"===typeof Reflect.metadata)return Reflect.metadata(c,a)},core_1=require("@angular/core"),integralui_core_1=require("./integralui.core"),integralui_common_service_1=require("../services/integralui.common.service"),integralui_data_service_1=require("../services/integralui.data.service"),integralui_menuitem_1=require("../components/integralui.menuitem"),IntegralUIMenu=
 function(c){function a(a,b,f,e,g,h){c.call(this,e);this.dataService=a;this.elemRef=b;this.elemRenderer=f;this.commonService=e;this.cmpResolver=g;this.baseService=h;this.trialRef=null;this.itemClick=new core_1.EventEmitter}__extends(a,c);a.prototype.ngOnInit=function(){this.baseService.setComponent(this);this.dataService.init([{data:this.items}]);this.generalClassName="iui-menu";this.initStyle()};a.prototype.ngAfterViewInit=function(){var a=this,b=setTimeout(function(){var d=a.cmpResolver.resolveComponentFactory(integralui_core_1.IntegralUITComponent);