+++ components/integralui.scrollbar.d.ts
@@ -1,0 +1,46 @@
+import { ElementRef, EventEmitter, Renderer } from '@angular/core';
+import { IntegralUIBaseComponent, IntegralUIOrientation } from './integralui.core';
+import { IntegralUICommonService } from '../services/integralui.common.service';
+export declare class IntegralUIScrollBar extends IntegralUIBaseComponent {
+    protected elemRef: ElementRef;
+    protected elemRenderer: Renderer;
+    protected commonService: IntegralUICommonService;
+    private currentMaxValue;
+    private currentMinValue;
+    private currentValue;
+    private currentSplitterDistance;
+    private emptySpace;
+    private maxPos;
+    private scrollSize;
+    private thumbSize;
+    private thumbPos;
+    private largeChange;
+    private smallChange;
+    height: number;
+    max: number;
+    min: number;
+    orientation: IntegralUIOrientation;
+    value: number;
+    width: number;
+    scrollStart: EventEmitter<any>;
+    scrollEnd: EventEmitter<any>;
+    valueChanged: EventEmitter<any>;
+    constructor(elemRef: ElementRef, elemRenderer: Renderer, commonService?: IntegralUICommonService);
+    ngOnInit(): void;
+    ngAfterViewInit(): void;
+    ngAfterContentInit(): void;
+    private isVertical();
+    getSize(): {
+        width: any;
+        height: any;
+    };
+    setScrollSize(value: any): void;
+    updateLayout(): void;
+    private thumbStartPos;
+    private isThumbActive;
+    private prevPos;
+    thumbMouseDown(e: any): void;
+    onWindowMouseMove(e: any): void;
+    onWindowMouseUp(e: any): void;
+    getScrollBarStyle(): any;
+}