+++ integralui.module.ts
@@ -1,27 +1,15 @@
-/*
-  filename: integralui.module.ts
-  version : 0.7.524 BETA
-  Copyright © 2016-2017 Lidor Systems. All rights reserved.
-
-  This file is part of the "IntegralUI Web" Library. 
-                                                                   
-  The contents of this file are subject to the IntegralUI Web License, and may not be used except in compliance with the License.
-  A copy of the License should have been installed in the product's root installation directory or it can be found at
-  http://www.lidorsystems.com/products/web/iui-web-license-agreement.pdf.
-                                                            
-  This SOFTWARE is provided "AS IS", WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the specific language 
-  governing rights and limitations under the License. Any infringement will be prosecuted under applicable laws.                           
-*/
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
 import { IntegralUIAccordion } from './components/integralui.accordion';
 import { IntegralUIAnchorStyle, IntegralUIDragWindow, IntegralUIFocus, IntegralUIHeaderItem, IntegralUIItem, IntegralUIObjectState, IntegralUISelectionMode, IntegralUITComponent, IntegralUITemplate, IntegralUIVisibility } from './components/integralui.core';
+import { IntegralUIBaseGrid, IntegralUIGridLines } from './components/integralui.base.grid';
 import { IntegralUIBaseList } from './components/integralui.base.list';
 import { IntegralUIComboBox } from './components/integralui.combobox';
 import { IntegralUIContextMenu, IntegralUIContextMenuComponent } from './directives/integralui.contextmenu';
 import { IntegralUIFrame } from './directives/integralui.frame';
+import { IntegralUIGrid } from './components/integralui.grid';
 import { IntegralUIGroupBox } from './components/integralui.groupbox';
 import { IntegralUIListBar } from './components/integralui.listbar';
 import { IntegralUIListBox } from './components/integralui.listbox';
@@ -32,12 +20,14 @@ import { IntegralUIMenu } from './components/integralui.menu';
 import { IntegralUIMenuItem } from './components/integralui.menuitem';
 import { IntegralUIPaginator } from './components/integralui.paginator';
 import { IntegralUIRange } from './directives/integralui.range';
+import { IntegralUIScrollBar } from './components/integralui.scrollbar';
 import { IntegralUISlideBar, IntegralUISlide } from './components/integralui.slidebar';
 import { IntegralUISplitContainer, IntegralUISplitContainerTags } from './components/integralui.splitcontainer';
 import { IntegralUISplitter } from './components/integralui.splitter';
 import { IntegralUITab } from './components/integralui.tab';
 import { IntegralUITabStrip } from './components/integralui.tabstrip';
 import { IntegralUITooltip, IntegralUITooltipComponent } from './directives/integralui.tooltip';
+import { IntegralUITreeGrid } from './components/integralui.treegrid';
 import { IntegralUITreeItem } from './components/integralui.treeitem';
 import { IntegralUITreeList, IntegralUITreeListItem } from './components/integralui.treelist';
 import { IntegralUITreeView } from './components/integralui.treeview';
@@ -49,6 +39,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
   	imports: [ CommonModule, FormsModule ],
   	declarations: [ 
   		IntegralUIAccordion,
+      IntegralUIBaseGrid,
       IntegralUIBaseList, 
       IntegralUIComboBox,
       IntegralUIContextMenu,
@@ -56,6 +47,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUIDragWindow,
       IntegralUIFocus,
       IntegralUIFrame,
+      IntegralUIGrid,
   		IntegralUIGroupBox, 
   		IntegralUIHeaderItem, 
   		IntegralUIItem, 
@@ -68,6 +60,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUIMenuItem, 
   		IntegralUIPaginator, 
       IntegralUIRange,
+      IntegralUIScrollBar,
   		IntegralUISlideBar, 
   		IntegralUISlide, 
       IntegralUISplitContainer, IntegralUISplitContainerTags,
@@ -78,6 +71,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUITemplate,
   		IntegralUITooltip, 
   		IntegralUITooltipComponent, 
+      IntegralUITreeGrid,
       IntegralUITreeItem, 
       IntegralUITreeList,
       IntegralUITreeListItem,
@@ -89,6 +83,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUIContextMenu,
       IntegralUIFocus, 
       IntegralUIFrame,
+      IntegralUIGrid,
   		IntegralUIGroupBox, 
   		IntegralUIHeaderItem, 
   		IntegralUIItem, 
@@ -101,6 +96,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUIMenuItem,
   		IntegralUIPaginator, 
       IntegralUIRange,
+      //IntegralUIScrollBar,
   		IntegralUISlideBar, 
   		IntegralUISlide, 
       IntegralUISplitContainer, IntegralUISplitContainerTags,
@@ -109,6 +105,7 @@ import { IntegralUIDragDropService } from './services/integralui.dragdrop.servic
       IntegralUITabStrip, 
       IntegralUITemplate,
   		IntegralUITooltip,
+      IntegralUITreeGrid,
       IntegralUITreeItem, 
       IntegralUITreeList,
       IntegralUITreeView